Modifies the relative "move" methods to take an alignment param for the adjacent axis.
That is, if you save `moveBelowSiblingLayer()` (vertical), you can
choose how the receiver aligns **horizontally** as well. By default, it
aligns to the leading edge, which is what you probably want most of the
time.

If you want the old behaviour (where it performed no such alignment),
pass in the `.none` option.